Rename app migrations
