+++ packages/server/src/api/controllers/application.ts
@@ -76,10 +76,10 @@ import { BASE_LAYOUT_PROP_IDS } from "../../constants/layouts"
 import sdk from "../../sdk"
 import { builderSocket } from "../../websockets"
 import { DefaultAppTheme, sdk as sharedCoreSDK } from "@budibase/shared-core"
-import * as appMigrations from "../../appMigrations"
+import * as appMigrations from "../../workspaceMigrations"
 import { createSampleDataTableScreen } from "../../constants/screens"
 import { defaultAppNavigator } from "../../constants/definitions"
-import { processMigrations } from "../../appMigrations/migrationsProcessor"
+import { processMigrations } from "../../workspaceMigrations/migrationsProcessor"
 
 // utility function, need to do away with this
 async function getLayouts() {
