+++ packages/server/src/middleware/workspaceMigrations.ts
@@ -1,23 +1,24 @@
 import { Header } from "@budibase/shared-core"
 import { UserCtx } from "@budibase/types"
 import type { Middleware, Next } from "koa"
-import { checkMissingMigrations } from "../appMigrations"
+import { checkMissingMigrations } from "../workspaceMigrations"
 import env from "../environment"
+import { context } from "@budibase/backend-core"
 
-export const appMigrations = (async (ctx: UserCtx, next: Next) => {
-  const { appId } = ctx
+export const workspaceMigrations = (async (ctx: UserCtx, next: Next) => {
+  const workspaceId = context.getWorkspaceId()
 
   // migrations can be disabled via environment variable if you
   // need to completely disable migrations, e.g. for testing
-  if (env.DISABLE_APP_MIGRATIONS) {
+  if (env.DISABLE_WORKSPACE_MIGRATIONS) {
     return next()
   }
 
-  if (!appId) {
+  if (!workspaceId) {
     return next()
   }
 
-  return checkMissingMigrations(ctx, next, appId)
+  return checkMissingMigrations(ctx, next, workspaceId)
 }) as Middleware
 
 export async function skipMigrationRedirect(ctx: UserCtx, next: Next) {
