+++ packages/server/src/startup/index.ts
@@ -16,7 +16,7 @@ import * as automations from "../automations"
 import * as fileSystem from "../utilities/fileSystem"
 import { default as eventEmitter, init as eventInit } from "../events"
 import * as bullboard from "../automations/bullboard"
-import * as appMigrations from "../appMigrations/queue"
+import * as appMigrations from "../workspaceMigrations/queue"
 import * as pro from "@budibase/pro"
 import * as api from "../api"
 import sdk from "../sdk"
